# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/ai_chat/chat/generate": {
            "post": {
                "description": "聊天接口(非流式)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai_chat模块"
                ],
                "summary": "聊天接口(非流式)",
                "parameters": [
                    {
                        "description": "聊天请求(非流式)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/aicode_internal_model_vo.AIChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/aicode_internal_common.BaseResponse-string"
                        }
                    }
                }
            }
        },
        "/ai_chat/chat/stream": {
            "post": {
                "description": "聊天接口(流式)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai_chat模块"
                ],
                "summary": "聊天接口(流式)",
                "parameters": [
                    {
                        "description": "聊天请求(流式)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/aicode_internal_model_vo.AIChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/aicode_internal_common.BaseResponse-string"
                        }
                    }
                }
            }
        },
        "/ai_code/gen": {
            "post": {
                "description": "代码生成",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai_code模块"
                ],
                "summary": "代码生成",
                "parameters": [
                    {
                        "description": "代码生成请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/aicode_internal_model_vo.AICodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/aicode_internal_common.BaseResponse-string"
                        }
                    }
                }
            }
        },
        "/ai_code/gen/stream": {
            "post": {
                "description": "代码生成流式",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai_code模块"
                ],
                "summary": "代码生成流式",
                "parameters": [
                    {
                        "description": "代码生成请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/aicode_internal_model_vo.AICodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/aicode_internal_common.BaseResponse-string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "aicode_internal_common.BaseResponse-string": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "aicode_internal_model_vo.AIChatMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "description": "\"user\" 或 \"assistant\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ChatRole"
                        }
                    ]
                }
            }
        },
        "aicode_internal_model_vo.AIChatRequest": {
            "type": "object",
            "required": [
                "model",
                "question"
            ],
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/aicode_internal_model_vo.AIChatMessage"
                    }
                },
                "model": {
                    "$ref": "#/definitions/consts.ChatModelType"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "aicode_internal_model_vo.AICodeMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "description": "\"user\" 或 \"assistant\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.ChatRole"
                        }
                    ]
                }
            }
        },
        "aicode_internal_model_vo.AICodeRequest": {
            "type": "object",
            "required": [
                "appId",
                "genType",
                "model",
                "question"
            ],
            "properties": {
                "appId": {
                    "type": "string"
                },
                "genType": {
                    "$ref": "#/definitions/consts.CodeGenarateType"
                },
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/aicode_internal_model_vo.AICodeMessage"
                    }
                },
                "model": {
                    "$ref": "#/definitions/consts.ChatModelType"
                },
                "question": {
                    "type": "string"
                }
            }
        },
...
        "consts.ChatModelType": {
            "type": "string",
            "enum": [
                "deepseek"
            ],
            "x-enum-varnames": [
                "ChatModelTypeDeepSeek"
            ]
        },
        "consts.ChatRole": {
            "type": "string",
            "enum": [
                "user",
                "assistant",
                "system"
            ],
            "x-enum-varnames": [
                "ChatRoleUser",
                "ChatRoleAssistant",
                "ChatRoleSystem"
            ]
        },
        "consts.CodeGenarateType": {
            "type": "string",
            "enum": [
                "single",
                "multi"
            ],
            "x-enum-varnames": [
                "CodeGenarateTypeSingle",
                "CodeGenarateTypeMulti"
            ]
        }
    }
}